Removed Devices.py and consolidated with Switch.py
