+++ setup.py
@@ -19,7 +19,7 @@ setup(name=NAME,
       author=AUTHOR,
       author_email=EMAIL,
       license='MIT',
-      packages=["capt", "capt.connector", "capt.function", "capt.procedure"],
+      packages=["capt", "capt.connector", "capt.function", "capt.procedure", "capt.utils"],
       install_requires=[
                   'urllib3>=1.13.1',
                   'requests>=2.9.1',
